Botones creados para llamar, whatsapp, email y Generador QR - Se dehabilito boton crear consultar en FireStore - Por ahora notificaciones no se han logrado implementar. Muy complicado

+++ app/src/main/java/com/bit/creciendojuntos/models/FCMResponse.java
@@ -3,14 +3,13 @@ package com.bit.creciendojuntos.models;
 import java.util.ArrayList;
 
 public class FCMResponse {
-
-    private long multicast_id;
+    private int  multicast_id;
     private int success;
     private int failure;
     private int canonical_ids;
     ArrayList<Object> results = new ArrayList<Object>();
 
-    public FCMResponse(long multicast_id, int success, int failure, int canonical_ids, ArrayList<Object> results) {
+    public FCMResponse(int multicast_id, int success, int failure, int canonical_ids, ArrayList<Object> results) {
         this.multicast_id = multicast_id;
         this.success = success;
         this.failure = failure;
@@ -18,11 +17,11 @@ public class FCMResponse {
         this.results = results;
     }
 
-    public long getMulticast_id() {
+    public int getMulticast_id() {
         return multicast_id;
     }
 
-    public void setMulticast_id(long multicast_id) {
+    public void setMulticast_id(int multicast_id) {
         this.multicast_id = multicast_id;
     }
